Add files via upload
Stock forecasting remains one of the most interesting topics of research. Understanding market behavior is crucial and is one of the most challenging problems in the financial world. Given the high dynamics of the market, it is not an easy task. Research in recent years has revealed interesting solutions based on various innovative ideas. Among them, hybrid machine learning models are of particular interest, as they can efficiently handle the problem. However, market behavior remains dependent on various external factors that are impossible to fully reveal, control, and integrate into the analysis. The goal of this project is to conduct in-depth research in this field using LSTM-GRU-based hybrid models, with one of the objectives being to integrate various technical indicators into the analysis.